backend/requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
google-genai>=1.0.0
chromadb>=0.5.20
python-dotenv>=1.0.1
pydantic>=2.10.0
httpx>=0.27.0
